# backend/src/api_server/models/expense.py
# coding: utf-8

"""
    Mouni Expense Manager API

    Simple API to manage group expenses and settle balances.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations

import json
import pprint
import re  # noqa: F401
from datetime import datetime
from typing import Any, ClassVar, Dict, List, Optional, Union

from pydantic import BaseModel, ConfigDict, Field, StrictStr
from typing_extensions import Annotated

from api_server.models.expense_involved import ExpenseInvolved

try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class Expense(BaseModel):
    """
    Expense
    """ # noqa: E501
    id: Optional[StrictStr] = None
    activity_id: Optional[StrictStr] = Field(default=None, alias="activityId")
    description: Annotated[str, Field(min_length=3, strict=True)]
    amount: Union[Annotated[float, Field(strict=True, ge=1)], Annotated[int, Field(strict=True, ge=1)]]
    currency: Annotated[str, Field(min_length=3, strict=True)]
    paid_by: StrictStr = Field(description="Member ID who paid", alias="paidBy")
    involved: List[ExpenseInvolved]
    var_date: datetime = Field(alias="date")
    created_at: Optional[datetime] = Field(default=None, description="Timestamp of when the resource was created", alias="createdAt")
    updated_at: Optional[datetime] = Field(default=None, description="Timestamp of the last update", alias="updatedAt")
    __properties: ClassVar[List[str]] = ["id", "activityId", "description", "amount", "currency", "paidBy", "involved", "date", "createdAt", "updatedAt"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of Expense from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        * OpenAPI `readOnly` fields are excluded.
        * OpenAPI `readOnly` fields are excluded.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
                "id",
                "activity_id",
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of each item in involved (list)
        _items = []
        if self.involved:
            for _item in self.involved:
                if _item:
                    _items.append(_item.to_dict())
            _dict['involved'] = _items
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of Expense from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "id": obj.get("id"),
            "activityId": obj.get("activityId"),
            "description": obj.get("description"),
            "amount": obj.get("amount"),
            "currency": obj.get("currency") if obj.get("currency") is not None else 'JPY',
            "paidBy": obj.get("paidBy"),
            "involved": [ExpenseInvolved.from_dict(_item) for _item in obj.get("involved")] if obj.get("involved") is not None else None,
            "date": obj.get("date"),
            "createdAt": obj.get("createdAt"),
            "updatedAt": obj.get("updatedAt")
        })
        return _obj