# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Mouni Expense Manager API
  description: Simple API to manage group expenses and settle balances.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /groups:
    post:
      tags:
        - groups
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
      responses:
        '201':
          description: Created group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
    get:
      tags:
        - groups
      summary: List all groups
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
  /groups/{groupId}:
    get:
      tags:
        - groups
      summary: Get group details
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Not found
    put:
      tags:
        - groups
      summary: Update a group
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdate'
      responses:
        '200':
          description: Updated group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Not found
    delete:
      tags:
        - groups
      summary: Delete a group
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Group deleted
        '404':
          description: Not found

  # Members CRUD
...
  /groups/{groupId}/activities/{activityId}/expenses:
    post:
      tags:
        - expenses
      summary: Add an expense to an activity
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: activityId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          description: Activity or member not found
    get:
      tags:
        - expenses
      summary: List expenses in an activity
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: activityId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the activity
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '404':
          description: Activity not found
  /groups/{groupId}/activities/{activityId}/expenses/{expenseId}:
    get:
      tags:
        - expenses
      summary: Get expense details
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: activityId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the activity
        - $ref: '#/components/parameters/ExpenseId'
      responses:
        '200':
          description: Expense details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          description: Activity or expense not found
    put:
      tags:
        - expenses
      summary: Update an expense in an activity
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: activityId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the activity
        - $ref: '#/components/parameters/ExpenseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseUpdate'
      responses:
        '200':
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          description: Activity or expense not found
    delete:
      tags:
        - expenses
      summary: Delete an expense
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: activityId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the activity
        - $ref: '#/components/parameters/ExpenseId'
      responses:
        '204':
          description: Expense deleted
        '404':
          description: Activity or expense not found

  # Settlements CRUD
...
components:
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
      description: The ID of the group
...
    ExpenseId:
      name: expenseId
      in: path
      required: true
      schema:
        type: string
      description: The ID of the expense
...
  schemas:
    Group:
      type: object
      properties:
        id:
          type: string
          example: "group123"
          readOnly: true
        name:
          type: string
          example: "Trip to Japan"
          minLength: 3
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
          readOnly: true
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the resource was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of the last update
      required:
        - name
    GroupCreate:
      type: object
      properties:
        name:
          type: string
          example: "Trip to Japan"
          minLength: 3
      required:
        - name
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
          example: "Trip to Japan"
          minLength: 3
      required:
        - name
    Member:
      type: object
      properties:
        id:
          type: string
          example: "user1"
          readOnly: true
        name:
          type: string
          example: "Alice"
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the resource was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of the last update
      required:
        - name
...
    Expense:
      type: object
      properties:
        id:
          type: string
          example: "expense1"
          readOnly: true
        activityId:
          type: string
          example: "activity1"
          readOnly: true
        description:
          type: string
          example: "Dinner at sushi place"
          minLength: 3
        amount:
          type: number
          format: float
          example: 120.50
          minimum: 1
        currency:
          type: string
          example: "JPY"
          minLength: 3
          default: "JPY"
        paidBy:
          type: string
          description: Member ID who paid
          example: "user1"
        involved:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseInvolved'
        date:
          type: string
          format: date-time
          example: "2025-08-11T18:30:00Z"
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the resource was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of the last update
      required:
        - description
        - amount
        - currency
        - paidBy
        - involved
        - date
    ExpenseCreate:
      type: object
      properties:
        description:
          type: string
          example: "Dinner at sushi place"
          minLength: 3
        amount:
          type: number
          format: float
          example: 120.50
          minimum: 1
        currency:
          type: string
          example: "JPY"
          minLength: 3
          enum:
          - JPY
          - USD
          - EUR
          - GBP
          - AUD
          - CAD
          - CHF
          - CNY
          - HKD
          - SGD
          default: "JPY"
        paidBy:
          type: string
          description: Member ID who paid
          example: "user1"
        involved:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseInvolved'
        equalSplit:
          type: boolean
          default: true
          description: >
            If true, ignore any shares provided in `involved` and split equally.
            If false, each involved entry must include a share.
        date:
          type: string
          format: date-time
          example: "2025-08-11T18:30:00Z"
      required:
        - description
        - amount
        - currency
        - paidBy
        - involved
        - date
    ExpenseInvolved:
      type: object
      required:
        - memberId
      properties:
        memberId:
          type: string
        share:
          type: number
          minimum: 0
          description: "How much this member owes. Required if equalSplit=false"
    ExpenseUpdate:
      type: object
      properties:
        description:
          type: string
          example: "Dinner at sushi place"
          minLength: 3
        amount:
          type: number
          format: float
          example: 120.50
          minimum: 1
        currency:
          type: string
          example: "JPY"
          minLength: 3
          enum:
          - JPY
          - USD
          - EUR
          - GBP
          - AUD
          - CAD
          - CHF
          - CNY
          - HKD
          - SGD
          default: "JPY"
        paidBy:
          type: string
          description: Member ID who paid
          example: "user1"
        involved:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseInvolved'
        equalSplit:
          type: boolean
          default: true
          description: >
            If true, ignore any shares provided in `involved` and split equally.
            If false, each involved entry must include a share.
        date:
          type: string
          format: date-time
          example: "2025-08-11T18:30:00Z"
      required:
        - description
        - amount
        - currency
        - paidBy
        - involved
        - date
...